+++ src/app/components/new-cluster-installation.js
@@ -26,6 +26,7 @@ import ClusterHostInfo from 'onepanel-gui/models/cluster-host-info';
 import PromiseObject from 'onedata-gui-common/utils/ember/promise-object';
 import watchTaskStatus from 'ember-onedata-onepanel-server/utils/watch-task-status';
 import I18n from 'onedata-gui-common/mixins/components/i18n';
+import safeExec from 'onedata-gui-common/utils/safe-method-execution';
 
 const {
   ProviderConfiguration,
@@ -42,11 +43,13 @@ function getClusterHostname(hostnames) {
 }
 
 function getDomain(hostname) {
-  return hostname.split('.').splice(1).join('.');
+  const segments = hostname.split('.');
+  return segments[segments.length - 1];
 }
 
 function getHostname(hostname) {
-  return hostname.split('.')[0];
+  const segments = hostname.split('.');
+  return segments.splice(0, segments.length - 1).join('.');
 }
 
 function getHostnamesOfType(hosts, type) {
@@ -121,6 +124,16 @@ export default Component.extend(I18n, {
   _hostTableValid: false,
 
   /**
+   * @type {string}
+   */
+  _newHostname: '',
+
+  /**
+   * @type {boolean}
+   */
+  _isSubmittingAddress: false,
+
+  /**
    * @type {function}
    */
   changeClusterName: undefined,
@@ -130,21 +143,22 @@ export default Component.extend(I18n, {
    */
   nextStep: undefined,
 
+  /**
+   * Changed on add new host button click
+   * @type {boolean}
+   */
+  addingNewHost: false,
+
   init() {
     this._super(...arguments);
 
+    this.observeResetNewHostname();
     this.changeCanDeploy();
     this.set(
       'hostsProxy',
       PromiseObject.create({
-        promise: new Promise((resolve, reject) => {
-          let gettingHosts = this.get('clusterManager').getHosts();
-          gettingHosts.then(hosts => {
-            hosts = A(hosts.map(h => ClusterHostInfo.create(h)));
-            resolve(hosts);
-          });
-          gettingHosts.catch(reject);
-        }),
+        promise: this.get('clusterManager').getHosts('known')
+          .then(hosts => A(hosts.map(h => ClusterHostInfo.create(h)))),
       })
     );
 
@@ -336,6 +350,14 @@ export default Component.extend(I18n, {
       scheduleOnce('afterRender', this, () => this.set('canDeploy', canDeploy));
     }),
 
+  observeResetNewHostname: observer('addingNewHost', function () {
+    if (this.get('addingNewHost')) {
+      scheduleOnce('afterRender', () => this.$('.input-add-host')[0].focus());
+    } else {
+      this.set('_newHostname', '');
+    }
+  }),
+
   actions: {
     zoneFormChanged(fieldName, value) {
       switch (fieldName) {
@@ -403,6 +425,24 @@ export default Component.extend(I18n, {
         this.get('globalNotify').backendError('deployment start', error);
       });
       return start;
+    },
+
+    submitNewHost() {
+      if (!this.get('_newHostname')) {
+        return Promise.reject();
+      } else {
+        const _newHostname = this.get('_newHostname');
+        this.set('_isSubmittingNewHost', true);
+        return this.get('clusterManager').addKnownHost(_newHostname)
+          .then(knownHost =>
+            this.get('hosts').pushObject(ClusterHostInfo.create(knownHost))
+          )
+          .then(() => safeExec(this, 'set', 'addingNewHost', false))
+          .catch(error =>
+            this.get('globalNotify').backendError(this.tt('addingNewHost'), error)
+          )
+          .finally(() => safeExec(this, 'set', '_isSubmittingNewHost', false));
+      }
     },
   },
 });