+++ src/app/locales/en/components/cluster-host-table.js
@@ -1,0 +1,9 @@
+export default {
+  headers: {
+    hosts: 'Hosts',
+    database: 'Database',
+    clusterWorker: 'Cluster Worker',
+    clusterManager: 'Cluster Manager',
+    primaryClusterManager: 'Primary Cluster Manager',
+  },
+};